cloud-run-backend/requirements.txt
# Research-driven podcast generation requirements
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP client
aiohttp>=3.8.0

# Google AI and Cloud services
google-genai
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.14.0
google-cloud-storage>=2.8.0
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.16.0
google-cloud-firestore>=2.11.0

# Research paper processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Audio processing
pydub>=0.25.1

# Image processing for thumbnails
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# System monitoring
psutil>=5.9.0

# OpenAI (fallback)
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0
